% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data.demografr_sims_abc}
\alias{combine_data.demografr_sims_abc}
\title{Combine multiple individual ABC simulation runs into one}
\usage{
\method{combine_data}{demografr_sims_abc}(...)
}
\arguments{
\item{...}{Either a list of objects of the class \code{demografr_sims} as produced
by the function \code{simulate_abc}, or individual objects of this class given
as standard function arguments, or paths to 'rds' files containing serializations of
such \code{demografr_sims} objects.}
}
\value{
A combined object of the class \code{demografr_sims}
}
\description{
Combine multiple individual ABC simulation runs into one
}
