% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grid.R
\name{combine_grid}
\alias{combine_grid}
\title{Combine multiple individual grid simulation runs into one}
\usage{
combine_grid(...)
}
\arguments{
\item{...}{Either a list of objects of the class \code{demografr_sims} as produced
by the function \code{simulate_abc}, or individual objects of this class given
as standard function arguments, or paths to 'rds' files containing serializations of
such \code{demografr_sims} objects.}
}
\value{
A combined object of the class \code{demografr_sims}
}
\description{
Combine multiple individual grid simulation runs into one
}
