% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fitness.R
\name{compute_fitness}
\alias{compute_fitness}
\title{Compute the fitness of the given parameter set against observed data}
\usage{
compute_fitness(
  params,
  model,
  functions,
  observed,
  sequence_length,
  recombination_rate,
  mutation_rate = 0,
  engine = NULL,
  model_args = NULL,
  engine_args = NULL,
  statistics = FALSE
)
}
\description{
Compute the fitness of the given parameter set against observed data
}
