% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Run cross-validation routines of the abc R package}
\usage{
cross_validate(x, nval, tols, method, ...)
}
\arguments{
\item{x}{A list of objects of the class \code{demografr_sims_abc} (each
representing an ABC inference result) for the function abc::cv4postpr,
or an object \code{demografr_abc.abc} for the function \code{abc::cv4abc}}

\item{nval, tols, method}{Required arguments of \code{abc::cv4postpr}}

\item{...}{Other optional arguments to be passed to \code{abc::cv4postpr}}
}
\value{
Object of the class \code{cv4postpr} and \code{demografr_cv}
}
\description{
This function is a convenience wrapper around the functions \code{cv4postpr}
and \code{anc::cv4abc} from the R package abc
}
\examples{
\dontshow{slendr::check_dependencies(python = TRUE, quit = TRUE) # dependencies must be present
}

#####################################################
# can we even distinguish between competing models?

\dontrun{
# read a list of three different ABC models
models <- lapply(c("X", "Y", "Z"), function(i) { readRDS(url(paste0(
"raw.githubusercontent.com/bodkan/demografr/refs/heads/main/inst/examples/downstream_abc",
i, ".rds"
))) })

# note that each element of the list is, indeed, a demografr ABC result
models[[1]]

 # run cross validation to find out if we have even the power to distinguish
# our competing models(see the abc package vignette for interpretation)
cv_models <- cross_validate(models, nval = 10, tols = c(0.005, 0.01, 0.05), method = "neuralnet")
cv_model

#####################################################
# can our model even estimate the parameters?

# read an example result of an ABC inference
abc_res <- readRDS(system.file("examples/basics_abc.rds", package = "demografr"))

# perform cross-validation
cv_params <- cross_validate(abc_res, nval = 10, tols = c(0.005, 0.01, 0.05),
                            method = "neuralnet")
cv_params
}

}
