% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model.R
\name{extract_model}
\alias{extract_model}
\title{Extract slendr model constructed from the ABC posterior distribution}
\usage{
extract_model(abc, summary = c("mode", "mean", "median"))
}
\arguments{
\item{abc}{ABC object generated by \code{perform_abc}}

\item{summary}{Which summary statistic of the posterior distribution to use? Options are
"mode" (i.e. MAP estimates), "mean", or "median".}
}
\description{
Extract slendr model constructed from the ABC posterior distribution
}
