% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_names.R
\name{extract_names}
\alias{extract_names}
\title{Extract names of individuals in a tree sequence}
\usage{
extract_names(ts, split = NULL)
}
\arguments{
\item{ts}{Tree-sequence object simulated by a slendr back-end engine}

\item{split}{Should sample names in the tree sequence be split by a column
(a population or time column)? Default is \code{NULL} and all names of
samples will be returned.}
}
\value{
A vector of character sample names. If \code{split} is specified,
  a list of such vectors is returned, one element of the list per population
  or sampling time.
}
\description{
Extract names of individuals in a tree sequence
}
