% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Extract inferred posterior(s) as a standard data frame}
\usage{
extract_posterior(abc, param = NULL, posterior = c("adj", "unadj"))
}
\arguments{
\item{abc}{An object produced by \code{perform_abc}}

\item{param}{A character vector containing either parameter names to summarize,
or a regex-like matches to be used for subsetting. If \code{NULL} (the default),
all parameters will be extracted.}

\item{posterior}{Should an "adj"-usted or "unadj"-usted posterior be extracted?
(Default is "adj").}
}
\value{
A data frame in the long format with posterior values of parameters
}
\description{
Extract inferred posterior(s) as a standard data frame
}
