% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract table of estimated model parameters}
\usage{
extract_summary(abc, type = NULL)
}
\arguments{
\item{abc}{ABC object generated by \code{perform_abc}}
}
\description{
Extract table of estimated model parameters
}
