% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Peform a test of the goodness-of-fit of a given model}
\usage{
goodness_of_fit(model, replicates, ...)
}
\arguments{
\item{replicates}{Number of replicates to be used to approximate the null distribution
of the goodness-of-fit statistic (passed as \code{nb.replicate} to the underlying
function \code{abc::gfit}))}

\item{...}{Other optional arguments to be passed to \code{abc::gfit}}

\item{models}{An object of the class \code{demografr_sims_abc} or
\code{demografr_abc.abc} produced by functions \code{simulate_abc} or
\code{run_anc}, respectively, which store simulated summary statistics needed
for the goodness-of-fit procedure.}
}
\value{
Object of the class \code{gfit} and \code{demografr_gfit}
}
\description{
This function is a convenience wrapper around the \code{gfit} from the
R package abc
}
