% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.demografr_abc.R
\name{hist.demografr_abc.abc}
\alias{hist.demografr_abc.abc}
\title{Plot histogram of posterior distribution(s)}
\usage{
\method{hist}{demografr_abc.abc}(x, param = NULL, ...)
}
\arguments{
\item{x}{ABC object generated by \code{perform_abc}}

\item{param}{A character vector containing either parameter names to summarize,
or a regex-like matches to be used for subsetting.}

\item{...}{Formal ellipsis argument of the \code{plot} method (unused)}
}
\description{
Plot histogram of posterior distribution(s)
}
