% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{perform_abc}
\alias{perform_abc}
\title{Perform ABC on data generated by \code{simulate_abc}}
\usage{
perform_abc(
  data,
  tolerance,
  method,
  hcorr = TRUE,
  transf = "none",
  logit.bounds,
  subset = NULL,
  kernel = "epanechnikov",
  numnet = 10,
  sizenet = 5,
  lambda = c(1e-04, 0.001, 0.01),
  race = FALSE,
  maxit = 500,
  ...
)
}
\arguments{
\item{data}{Simulated data set produced by \code{simulate_abc}}

\item{tolerance}{The proportion of simulated samples to accept}

\item{method}{
    a character string indicating the type of ABC algorithm to be
    applied.  Possible values are \code{"rejection"},
    \code{"loclinear"}, \code{"neuralnet"} and \code{"ridge"}. See also
    \code{Details}.}

\item{hcorr}{
    logical, the conditional heteroscedastic model is applied if
    \code{TRUE} (default).}

\item{transf}{
    a vector of character strings indicating the kind of transformation
    to be applied to the parameter values. The possible values are
    \code{"log"}, \code{"logit"}, and \code{"none"} (default), when no
    is transformation applied. See also \code{Details}.}

\item{logit.bounds}{
    a matrix of bounds if \code{transf} is \code{"logit"}. 
    The matrix has as many lines as parameters (including the ones 
    that are not \code{"logit"} transformed) and 2 columns. First column is the 
    minimum bound and second column is the maximum bound.}

\item{subset}{
    a logical expression indicating elements or rows to keep. Missing
    values in \code{param} and/or \code{sumstat} are taken as
    \code{FALSE}.}

\item{kernel}{
    a character string specifying the kernel to be used when
    \code{method} is \code{"loclinear"}, \code{"neuralnet"} or \code{"ridge"}. Defaults
    to \code{"epanechnikov"}. See \code{\link{density}} for details.}

\item{numnet}{
    the number of neural networks when \code{method} is
    \code{"neuralnet"}. Defaults to 10. It indicates the number of times
    the function \code{\link{nnet}} is called.}

\item{sizenet}{
    the number of units in the hidden layer. Defaults to 5. Can be zero
    if there are no skip-layer units. See \code{\link{nnet}} for more
    details.}

\item{lambda}{
    a numeric vector or a single value indicating the weight decay when
    \code{method} is \code{"neuralnet"}. See \code{\link{nnet}} for more
    details. By default, 0.0001, 0.001, or 0.01 is randomly chosen for
    each of the networks.}

\item{maxit}{
    numeric, the maximum number of iterations. Defaults to 500. Applies
    only when \code{method} is \code{"neuralnet"}. See also
    \code{\link{nnet}}.}

\item{...}{
    other arguments passed to \code{\link{nnet}}.}
}
\description{
Runs a selected ABC method on simulated data using R package abc as
an inference engine.
}
\details{
This function serves as a wrapper around the function \code{\link[abc]{abc}} from
the R package abc. All function arguments except to \code{data} are passed to the
\code{\link[abc]{abc}} function, appropriately unpacking the prior sample matrix,
and binding together matrices with observed statistics and simulated statistics
in the format required by the inference function.

This function exists to avoid the need to manually track parameter matrices
and summary statistics as inputs to the \code{\link[abc]{abc}} function but acts
entirely transparently. A such, all implementation details can be found in the abc
vignette and the manpage which you can access by typing \code{?abc::abc}.
}
