% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_cv.R
\name{perform_cv}
\alias{perform_cv}
\title{Perform cross-validation using the \code{anc::cv4postpr} function}
\usage{
perform_cv(models, nval, tols, method, ...)
}
\arguments{
\item{models}{A list of objects of the class \code{demografr_sims_abc} or
\code{demografr_abc.abc} which store simulated summary statistics needed for
calling \code{cv4postpr}}

\item{nval, tols, method}{Required arguments of \code{cv4postpr}}
}
\value{
Object of the class \code{cv4postpr} and \code{demografr_cv}
}
\description{
This function is a thin, convenience wrapper around the \code{cv4postpr} from the
R package abc
}
