% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.demografr_abc.R
\name{plot.demografr_abc.abc}
\alias{plot.demografr_abc.abc}
\title{Plot diagnostics of posterior distribution(s)}
\usage{
\method{plot}{demografr_abc.abc}(x, param = NULL, ...)
}
\description{
Plot diagnostics of posterior distribution(s)
}
