% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.demografr_abc_sims}
\alias{print.demografr_abc_sims}
\title{Print a brief summary of a set of demografr simulations}
\usage{
\method{print}{demografr_abc_sims}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{demografr_abc_sims}, generated
by \code{simulate_abc}}
}
\value{
Used exclusively for printing
}
\description{
Print a brief summary of a set of demografr simulations
}
\keyword{internal}
