% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.demografr_cv.R
\name{print.demografr_cv}
\alias{print.demografr_cv}
\title{Print a brief summary of the result of the \code{abc::cv4postpr} function}
\usage{
\method{print}{demografr_cv}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{demografr_cv}, generated by \code{perform_cv}}
}
\value{
Used exclusively for printing
}
\description{
Print a brief summary of the result of the \code{abc::cv4postpr} function
}
