% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.demografr_postpr}
\alias{print.demografr_postpr}
\title{Print a brief summary of the result of the \code{abc::postpr} function}
\usage{
\method{print}{demografr_postpr}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{demografr_postpr}, generated by \code{model_selection}}
}
\value{
Used exclusively for printing
}
\description{
Print a brief summary of the result of the \code{abc::postpr} function
}
\keyword{internal}
