% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.demografr_sim.R
\name{print.demografr_sims.abc}
\alias{print.demografr_sims.abc}
\title{Print a brief summary of a set of demografr simulations}
\usage{
\method{print}{demografr_sims.abc}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{demografr_sims}, generated by \code{simulate_abc}}
}
\value{
Used exclusively for printing
}
\description{
Print a brief summary of a set of demografr simulations
}
\keyword{internal}
