% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated-models.R
\name{random_model}
\alias{random_model}
\title{Create random model}
\usage{
random_model(
  n_populations,
  population_size,
  n_gene_flow,
  rate_gene_flow = c(0.01, 0.99),
  time_span
)
}
\arguments{
\item{n_populations}{The desired number of populations for the model}

\item{population_size}{Integer number defining the populations size of all
populations}

\item{n_gene_flow}{Integer number defining the number of gene_flow events}

\item{rate_gene_flow}{Vector that specifies the min and max gene flow rate}

\item{time_span}{Integer number defining on how long the simulation
of the populations will last, to scale the edge lengths of the tree
accordingly}
}
\value{
A slendr model
}
\description{
Creates a slendr model based on a given number of populations and a given
number of gene flow events. A random ape tree is created according to the
specified number of populations. With this tree, the function
\code{tree_model} is called to create the model.
}
\examples{
tree_model(n_populations = 4, population_size = 1000, n_gene_flow = 3,
  rate_gene_flow = 0.5, time_span = 100)
tree_model(n_populations = 6, population_size = 200, n_gene_flow = 4,
  rate_gene_flow = c(0.2, 0.9), time_span = 1000)
}
