% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated-models.R
\name{random_populations}
\alias{random_populations}
\title{Create a list of a given number of random slendr populations}
\usage{
random_populations(n, time_span, Ne = 10000, names = NULL)
}
\arguments{
\item{n}{The desired number of populations that should be created}

\item{time_span}{Integer number defining on how long the simulation
of the populations will last, to scale the edge lengths of the tree
accordingly}

\item{Ne}{Integer number defining the effective populations size of all populations}

\item{names}{Names of the populations created. If \code{NULL}, a set of names
"pop1", "pop2", ... will be created.}
}
\value{
A set of slendr populations
}
\description{
Creates a random list of slendr populations of a given length. Internally, a
random phylogenetic tree of a given size is created and \code{tree_populations} is
called on that tree to create the populations themselves.
}
\examples{
random_populations(5, time_span = 10000, population_size = 1000)
}
