% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_env.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{init_env}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{slendr}{\code{\link[slendr]{init_env}}}
}}

