% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_epochs.R
\name{resize_epochs}
\alias{resize_epochs}
\title{Plan resize events for the given population}
\usage{
resize_epochs(pop, epochs, sizes, until = NULL, times = NULL)
}
\arguments{
\item{pop}{A slendr population object}

\item{epochs}{The number of resize epochs to schedule}

\item{sizes}{A vector of population sizes (one element of the vector for each epoch)}

\item{until}{If \code{times} is not given, this argument gives the time of the last resize event.
See \code{times}.}

\item{times}{A vector of times of each resize event (one element of the vector for each epoch).
If \code{NULL}, the times are computed by dividing the time between the last given slendr
event in the population's history and the last resize time \code{until}.}
}
\description{
Plan resize events for the given population
}
