% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_names}
\alias{sample_names}
\title{Extract names of individuals in a tree sequence}
\usage{
sample_names(ts, split = FALSE)
}
\arguments{
\item{ts}{Tree-sequence object simulated by a slendr back-end engine}

\item{split}{Should sample names be split by population? Default is \code{FALSE}}
}
\value{
A vector of character sample names. If \code{split} is \code{TRUE},
  a list of such vectors is returned, one element of the list per population.
}
\description{
Extract names of individuals in a tree sequence
}
