% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model.R
\name{select_model}
\alias{select_model}
\title{Perform selection between different ABC models}
\usage{
select_model(models, ...)
}
\arguments{
\item{models}{A list of objects of the class \code{demografr_sims_abc} or
\code{demografr_abc.abc} which store simulated summary statistics needed for
calling \code{abc::postpr}}

\item{...}{Other optional arguments to be passed to \code{abc::postpr}}

\item{tol, method}{Required arguments of \code{abc::postpr}}
}
\value{
Object of the class \code{postpr} and \code{demografr_postpr}
}
\description{
This function internally uses \code{abc::postpr} for computing posterior probabilities
of models, Bayes factors, etc., in order to aid in the selection of the most likely
model.
}
