% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.R
\name{simulate_model}
\alias{simulate_model}
\title{Simulate a single tree sequence from the given inference setup}
\usage{
simulate_model(
  model,
  parameters,
  sequence_length,
  recombination_rate,
  mutation_rate = 0,
  data = NULL,
  format = c("ts", "files"),
  engine = NULL,
  model_args = NULL,
  engine_args = NULL,
  attempts = 1000
)
}
\arguments{
\item{model}{Either a slendr model generating function (in which case \code{engine} must
be either "msprime" or "slim", i.e. one of the two of slendr's simulation back ends),
or a path to a custom user-defined SLiM or msprime script (in which case \code{engine}
must be "files").}

\item{parameters}{A list of prior distributions to use for sampling of model parameters}

\item{sequence_length}{Amount of sequence to simulate using slendr (in base pairs).
Ignored when custom simulations scripts are provided.}

\item{recombination_rate}{Recombination rate to use for the simulation
Ignored when custom simulations scripts are provided.}

\item{mutation_rate}{Mutation rate to use for the simulation
Ignored when custom simulations scripts are provided.}

\item{data}{A named list of data-generating functions. The names then represent all possible
arguments of simulated summary statistic functions.}

\item{format}{In which format will the model generate results to be used for computing
simulated summary statistics?}

\item{engine}{Which simulation engine to use? Values "msprime" and "slim" will use one of
the built-in slendr simulation back ends. Which engine will be used is determined
by the nature of the \code{model}. If \code{engine = NULL}, then spatial slendr models will
by default use the "slim" back end, non-spatial models will use the "msprime" back end, and
custom user-defined model scripts will use the "custom" engine. Setting this argument
explicitly will change the back ends (where appropriate). Setting this argument for custom
simulation script has no effect.}

\item{model_args}{Optional (non-prior) arguments for the slendr model generating function.
Setting this argument for custom simulation script has no effect.}

\item{engine_args}{Optional arguments for the slendr simulation back end. Setting this
argument for custom simulation script has no effect.}

\item{attempts}{Maximum number of attempts to generate prior values for a valid demographic
model (i.e. model which generates an output without error, default is 1000)}
}
\value{
Either a tree-sequence object of the class \code{slendr_ts} when a slendr model
  was simulated, or a path to an output file when a custom simulation script was used.
}
\description{
Simulates a tree sequence object from a model with parameters as specified either
by sampling from priors or by a given list of fixed parameter values
}
\details{
This function is useful to generate a small tree sequence to be used when developing
summary statistic functions inference using demografr.
}
