% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ts.R
\name{simulate_output}
\alias{simulate_output}
\title{Simulate a single tree sequence from the given inference setup}
\usage{
simulate_output(
  model,
  parameters,
  sequence_length,
  recombination_rate,
  mutation_rate = 0,
  attempts = 1000,
  engine = NULL,
  model_args = NULL,
  engine_args = NULL,
  random_seed = NULL
)
}
\arguments{
\item{model}{Either a slendr model generating function (in which case \code{engine} must
be either "msprime" or "slim", i.e. one of the two of slendr's simulation back ends),
or a path to a custom user-defined SLiM or msprime script (in which case \code{engine}
must be "custom").}

\item{parameters}{A list of prior distributions to use for sampling of model parameters}

\item{sequence_length}{Amount of sequence to simulate using slendr (in numbers of basepairs)}

\item{recombination_rate}{Recombination rate to use for the simulation}

\item{mutation_rate}{Mutation rate to use for the simulation}

\item{attempts}{Maximum number of attempts to generate prior values for a valid demographic
model (default is 1000)}

\item{engine}{Which simulation engine to use? Values "msprime" and "slim" will use one of
the built-in slendr simulation back ends. Which engine will be used is determined
by the nature of the \code{model}. If \code{engine = NULL}, then spatial slendr models will
by default use the "slim" back end, non-spatial models will use the "msprime" back end, and
custom user-defined model scripts will use the "custom" engine. Setting this argument
explicitly will change the back ends (where appropriate). Setting this argument for custom
simulation script has no effect.}

\item{model_args}{Optional (non-prior) arguments for the slendr model generating function.
Setting this argument for custom simulation script has no effect.}

\item{engine_args}{Optional arguments for the slendr simulation back end. Setting this
argument for custom simulation script has no effect.}

\item{random_seed}{Random seed to be used for simulation and (potentially) adding of mutations
to a simulated tree sequence}
}
\value{
A tree-sequence object of the class \code{slendr_ts}
}
\description{
Simulates a tree sequence object from a model with parameters as specified either
by sampling from priors or by a given list of fixed parameter values
}
\details{
This function is useful to generate a small tree sequence to be used when developing
summary statistic functions inference using demografr.
}
