% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{simulate_ts}
\alias{simulate_ts}
\title{Simulate a single tree sequence from the given ABC setup}
\usage{
simulate_ts(
  model,
  priors,
  sequence_length = 1e+06,
  recombination_rate = 0,
  mutation_rate = 0,
  samples = NULL,
  engine = NULL,
  model_args = NULL,
  engine_args = NULL
)
}
\arguments{
\item{model}{A compiled slendr model object}

\item{priors}{A list of prior distributions to use for sampling of model parameters}

\item{sequence_length}{Amount of sequence to simulate using slendr (in numbers of basepairs)}

\item{recombination_rate}{Recombination rate to use for the simulation}

\item{mutation_rate}{Mutation rate to use for the simulation}

\item{model_args}{Optional (non-prior) arguments for the scaffold model generating function}

\item{engine_args}{Optional arguments for the slendr simulation back ends}
}
\description{
Simulates a tree sequence object from a model with parameters sampled from priors
}
\details{
This function is useful to generate a small tree sequence to be used when developing
summary statistic functions for ABC inference using demografr.
}
