% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.demografr_cv.R
\name{summary.demografr_cv}
\alias{summary.demografr_cv}
\title{Summary results of the \code{cv4postpr} procedure from the abc R package}
\usage{
\method{summary}{demografr_cv}(object, ...)
}
\arguments{
\item{object}{ABC object generated by \code{run_abc}}

\item{...}{Formal ellipsis argument of the \code{summary} method (unused)}
}
\value{
...
}
\description{
Summary results of the \code{cv4postpr} procedure from the abc R package
}
\keyword{internal}
