% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated-models.R
\name{tree_model}
\alias{tree_model}
\title{Create a slendr model based on a given phylogenetic tree}
\usage{
tree_model(tree, time_span, N = 1000, generation_time = 1)
}
\arguments{
\item{tree}{A phylogenetic tree of the class \code{phylo} (see the ape R package
for more details on the format)}

\item{time_span}{Integer number specifying what time span should the encoded demographic
history cover.}

\item{N}{Effective population size of all lineages. A value of 1000 is used by default.}

\item{generation_time}{Generation time used by the model}

\item{gene_flows}{Integer number defining the number of gene-flow events}

\item{rate}{Vector that specifies the min and max gene flow rate}
}
\value{
An object of the \code{slendr} class
}
\description{
Creates a slendr model based on a given phylogenetic tree and a given number
of gene flow events. First, the function \code{tree_populations} is used to get
generate a slendr model capturing a demographic history of a aset of populations
based on a given phylogenetic tree.
Gene flow events between these populations are created randomly. The rate of
the gene flow events can either be a fixed value or sampled in a specific range.
}
\examples{
tree <- ape::rtree(4)
tree_model(tree = tree, population_size = 1000, time_span = 100)
}
