% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated-models.R
\name{tree_populations}
\alias{tree_populations}
\title{Create a list of slendr populations based on given phylogenetic tree}
\usage{
tree_populations(tree, time_span, N = 1000, verbose = FALSE)
}
\arguments{
\item{tree}{A phylogenetic tree of the class \code{phylo} (see the ape R package
for more details on the format)}

\item{time_span}{Integer number specifying what time span should the encoded demographic
history cover.}

\item{N}{Effective population size of all lineages. A value of 1000 is used by default.}
}
\value{
A list of populations of the class \code{slendr}
}
\description{
Create a list of slendr populations based on given phylogenetic tree
}
\examples{
# create a random phylogenetic tree
tree <- ape::rtopology(3, tip.label = sprintf("pop_\%d", 1:3), rooted = TRUE)
populations <- tree_populations(tree = tree, time_span = 1000)
}
