% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Perform cross-validation using the \code{anc::cv4postpr} function}
\usage{
cross_validate(x, nval, tols, method, ...)
}
\arguments{
\item{x}{A list of objects of the class \code{demografr_sims_abc} (each
representing a model for model selection via the function abc::cv4postpr),
or a single object \code{demografr_abc.abc} to perform cross-validation
by calling the function \code{abc::cv4abc}.}

\item{nval, tols, method}{Required arguments of \code{abc::cv4postpr}}

\item{...}{Other optional arguments to be passed to \code{abc::cv4postpr}}
}
\value{
Object of the class \code{cv4postpr} and \code{demografr_cv}
}
\description{
This function is a convenience wrapper around the \code{cv4postpr} from the
R package abc
}
\examples{
\dontshow{slendr::check_dependencies(python = TRUE, quit = TRUE) # dependencies must be present
}

########################################
# cross-validation

# read an example result of an ABC inference
abc_res <- readRDS(system.file("examples/basics_abc.rds", package = "demografr"))
# perform cross-validation
cv_res <- cross_validate(abc_res, nval = 10, tols = c(0.005, 0.01, 0.05), method = "neuralnet")
cv_res

########################################
# model selection

# \dontrun{
# read a list of three different ABC models
models <- lapply(c("X", "Y", "Z"), function(i) {
  readRDS(
    url(
      sprintf("https://raw.githubusercontent.com/bodkan/demografr/refs/heads/main/inst/examples/downstream_abc\%s.rds", i)
    )
  )
})
# note that each element of the list is, indeed, a demografr ABC result
models[[1]]
mod_sel <- cross_validate(models, nval = 10, tols = c(0.005, 0.01, 0.05), method = "neuralnet")
mod_sel
}

}
