% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Extract inferred posterior(s) as a standard data frame}
\usage{
extract_posterior(abc, posterior = c("adj", "unadj"))
}
\description{
Extract inferred posterior(s) as a standard data frame
}
