% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract table of estimated model parameters}
\usage{
extract_summary(abc, param = NULL)
}
\arguments{
\item{abc}{ABC object generated by \code{perform_abc}}

\item{param}{A character vector containing either parameter names to summarize,
or a regex-like matches to be used for subsetting.}
}
\description{
Extract table of estimated model parameters
}
