% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract table of estimated model parameters}
\usage{
extract_summary(abc, param = NULL, type = NULL)
}
\arguments{
\item{abc}{ABC object generated by \code{perform_abc}}

\item{param}{Name of the model parameter}

\item{type}{Type of the model parameter ("Ne", "Tsplit", etc.)}
}
\description{
Extract table of estimated model parameters
}
