% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract table of estimated model parameters}
\usage{
extract_summary(abc, param = NULL)
}
\arguments{
\item{abc}{ABC object generated by \code{run_abc}}

\item{param}{A character vector containing either parameter names to summarize,
or a regex-like matches to be used for subsetting. If \code{NULL} (the default),
all parameters will be extracted.}
}
\value{
A data frame object with posterior summary statistics
}
\description{
Extract table of estimated model parameters
}
\examples{
# read example ABC result with an inferred joint posterior distribution
abc_res <- readRDS(system.file("examples/basics_abc.rds", package = "demografr"))

extract_summary(abc_res)

}
