% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hist.demografr_abc}
\alias{hist.demografr_abc}
\title{Plot histogram of posterior distribution(s)}
\usage{
\method{hist}{demografr_abc}(x, param = NULL, ...)
}
\description{
Plot histogram of posterior distribution(s)
}
