% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{perform_abc}
\alias{perform_abc}
\title{Perform ABC inference on the data generated by \code{simulate_abc}}
\usage{
perform_abc(data, tolerance, method, ...)
}
\arguments{
\item{tolerance}{Proportion of samples to accept in the neighborhood of observed values}

\item{method}{ABC algorithm to use for inference. Options are "rejection", "loclinear",
"neuralnet", and "ridge". For details, see the \code{abc} function from the package abc.}
}
\description{
Perform ABC inference on the data generated by \code{simulate_abc}
}
