% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_abc.R
\name{perform_abc}
\alias{perform_abc}
\title{Perform ABC on data generated by \code{simulate_abc}}
\usage{
perform_abc(data, engine, ...)
}
\arguments{
\item{data}{Simulated data set produced by \code{simulate_abc}}

\item{engine}{Which ABC engine to use? As of the current version of demografr, the
valid choices are either "abc" or "ABC_mcmc".}

\item{...}{Additional arguments passed on the the \code{abc} function from
the abc package}
}
\description{
Runs a selected ABC method on simulated data using R package abc as
an inference engine.
}
\details{
This function serves as a wrapper around the function \code{\link[abc]{abc}} from
the R package abc. All function arguments except to \code{data} are passed to the
\code{\link[abc]{abc}} function, appropriately unpacking the prior sample matrix,
and binding together matrices with observed statistics and simulated statistics
in the format required by the inference function.

This function exists to avoid the need to manually track parameter matrices
and summary statistics as inputs to the \code{\link[abc]{abc}} function but acts
entirely transparently. A such, all implementation details can be found in the abc
vignette and the manpage which you can access by typing \code{?abc::abc}.
}
