% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.demografr_abc.R
\name{plot.demografr_abc.abc}
\alias{plot.demografr_abc.abc}
\title{Plot diagnostics of posterior distribution(s)}
\usage{
\method{plot}{demografr_abc.abc}(x, param = NULL, ...)
}
\arguments{
\item{x}{ABC object generated by \code{perform_abc}}

\item{param}{A character vector containing either parameter names to summarize,
or a regex-like matches to be used for subsetting. If \code{NULL} (the default),
all parameters will be extracted.}

\item{...}{Formal ellipsis argument of the \code{plot} method (unused)}
}
\value{
Used exclusively for printing
}
\description{
Plot diagnostics of posterior distribution(s)
}
