% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.demografr_abc.R
\name{plot.demografr_abc.abc}
\alias{plot.demografr_abc.abc}
\title{Plot diagnostics of posterior distribution(s)}
\usage{
\method{plot}{demografr_abc.abc}(x, param = NULL, ...)
}
\arguments{
\item{x}{ABC object generated by \code{run_abc}}

\item{param}{A character vector containing either parameter names to summarize,
or a regex-like matches to be used for subsetting. If \code{NULL} (the default),
all parameters will be extracted.}

\item{...}{Formal ellipsis argument of the \code{plot} method (unused)}
}
\value{
Used exclusively for printing
}
\description{
Plot diagnostics of posterior distribution(s)
}
\examples{
# read inferred posterior distribution object from an example ABC run
abc_res <- readRDS(system.file("examples/basics_abc.rds", package = "demografr"))

# plot diagnostics for all parameters (in sequence)
# plot(abc_res)

# plot diagnostics only for selected parameters
plot(abc_res, param = "^Ne_")
plot(abc_res, param = "gf_BC")

}
