% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.demografr_cv.R
\name{plot.demografr_cv}
\alias{plot.demografr_cv}
\title{Plot results of the cv4postpr procedure from the abc R package}
\usage{
\method{plot}{demografr_cv}(x, file = NULL, ...)
}
\arguments{
\item{x}{ABC object generated by \code{run_abc}}

\item{file}{Path to a PDF file for an output figure (if \code{NULL}, which is the default,
the figure will not be saved to a file)}

\item{...}{Formal ellipsis argument of the \code{plot} method}
}
\value{
Used exclusively for printing
}
\description{
Plot results of the cv4postpr procedure from the abc R package
}
