% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior distribution of given parameters}
\usage{
plot_posterior(
  abc,
  param = NULL,
  posterior = c("adj", "unadj"),
  summary = c("mode", "mean", "median"),
  geom = ggplot2::geom_density,
  facets = TRUE,
  xlim = NULL,
  ...
)
}
\arguments{
\item{abc}{ABC object generated by \code{perform_abc}}

\item{param}{A character vector containing either parameter names to summarize,
or a regex-like matches to be used for subsetting.}
}
\description{
Plot posterior distribution of given parameters
}
