% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot the results of a posterior predictive check for a given summary statistic}
\usage{
plot_prediction(data, stat, file = NULL, facets = TRUE, ...)
}
\arguments{
\item{data}{Data frame with predictions generated from ABC posteriors, as generated
by the \code{predict} function}

\item{stat}{Which tree-sequence summary statistic to extract? If \code{NULL}, all
summary statistics will be visualized.}

\item{file}{Output path to a PDF figure to be saved}

\item{facets}{Should each summary statistic be plotted on its own facet (default is
\code{TRUE})?}

\item{...}{Optional argument which will be passed to \code{ggsave}}
}
\value{
If \code{file = NULL}, a ggplot2 object (or a list of such objects, if
  multiple summary statistics are visualized), otherwise nothing is returned and the
  plots are instead saved to a file.
}
\description{
Plot the results of a posterior predictive check for a given summary statistic
}
