% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.demografr_abc.abc.R
\name{predict.demografr_abc.abc}
\alias{predict.demografr_abc.abc}
\title{Generate summary statistics from the inferred posterior distribution of parameters}
\usage{
\method{predict}{demografr_abc.abc}(
  abc,
  samples,
  stat = NULL,
  posterior = c("unadj", "adj"),
  functions = NULL
)
}
\arguments{
\item{abc}{ABC object generated by \code{perform_abc}}

\item{samples}{Number of draws from the posterior distribution to simulate}

\item{stat}{Which tree-sequence summary statistic to compute? If \code{NULL}, all summary
statistics will be computed.}

\item{posterior}{Should 'unadj'usted or 'adj'usted parameters be sampled? Default
(recommended) value is 'unadj'usted.}

\item{functions}{A named list of summary statistic tree-sequence functions to be
applied to each simulated tree sequence. If \code{NULL} (the default), the same
summary statistics will be computed as those that were used in the ABC inference
itself. Otherwise, custom tree-sequence summary statistics can be provided.}
}
\value{
A data frame object with the results of posterior simulations, with values of
 each summary statistic stored in a list-column (i.e. the same format as used by
 the function \code{simulate_grid}, as the predictions are generated internally using
 this function)
}
\description{
The function draws parameter values from the posterior distribution and then
internally runs the function \code{simulate_grid} to run simulations from
the sampled parameters and compute tree-sequence summary statistics
}
