% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_script.R
\name{run_script}
\alias{run_script}
\title{Run a custom tailored SLiM or Python (msprime) simulation script}
\usage{
run_script(script, ...)
}
\arguments{
\item{script}{Path to the SLiM or Python script}

\item{...}{Model arguments which will be provided to the script on the command line}
}
\description{
This function will take as input a path to a SLiM or Python script (detecting
which one of the two it is based on the script's contents) and run it on the
command line with model arguments formatted on the command line automatically.
}
