% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{sample_prior}
\alias{sample_prior}
\title{Sample value from a given prior sampling formula object}
\usage{
sample_prior(f)
}
\arguments{
\item{f}{A formula object of a form <variable> ~ <function>(parameters)}
}
\description{
Sample value from a given prior sampling formula object
}
