% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{sample_prior}
\alias{sample_prior}
\title{Sample value from a given prior sampling formula object}
\usage{
sample_prior(f)
}
\arguments{
\item{f}{Formula-based prior sampling expression such as <variable> ~ <sampling statement>}
}
\value{
A list of two elements, "variable" containing the name of the sampled variable,
  and "value" containing the actual value of the sampled prior.
}
\description{
Sample value from a given prior sampling formula object
}
