% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{simulate_abc}
\alias{simulate_abc}
\title{Simulate data for ABC inference using specified priors}
\usage{
simulate_abc(
  model,
  priors,
  summary_funs,
  observed,
  iterations = 1,
  epochs = 1,
  mutation_rate = 0,
  sequence_length = 1e+07,
  recombination_rate = 1e-08
)
}
\arguments{
\item{model}{A compiled slendr model object}

\item{priors}{A list of prior distributions to use for sampling of model parameters}

\item{summary_funs}{A named list of summary statistic functions to apply on simulated
tree sequences}

\item{observed}{A named list of observed summary statistics}
}
\description{
Simulate data for ABC inference using specified priors
}
