% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_abc.R
\name{simulate_abc}
\alias{simulate_abc}
\title{Simulate data for ABC inference using specified priors}
\usage{
simulate_abc(
  model,
  priors,
  functions,
  observed,
  iterations,
  sequence_length,
  recombination_rate,
  mutation_rate = 0,
  model_args = NULL,
  engine_args = NULL,
  packages = NULL,
  engine = c("msprime", "slim"),
  debug = FALSE,
  attempts = 1000
)
}
\arguments{
\item{model}{A slendr model generating function}

\item{priors}{A list of prior distributions to use for sampling of model parameters}

\item{functions}{A named list of summary statistic functions to apply on simulated
tree sequences}

\item{observed}{A named list of observed summary statistics}

\item{iterations}{How many simulation replicates to run?}

\item{sequence_length}{Amount of sequence to simulate using slendr (in numbers of basepairs)}

\item{recombination_rate}{Recombination rate to use for the simulation}

\item{mutation_rate}{Mutation rate to use for the simulation}

\item{model_args}{Optional (non-prior) arguments for the scaffold model generating function}

\item{engine_args}{Optional arguments for the slendr simulation back ends}

\item{packages}{A character vector with package names used by user-defined summary statistic
functions. Only relevant when \code{future::plan("multisession", ...)} was initialized.}

\item{engine}{Which simulation engine to use? Values "msprime" and "slendr::"
will use the built-in slendr simulation back ends.}

\item{debug}{Only perform a single ABC simulation run, skipping parallelization}

\item{attempts}{Maximum number of attempts to generate prior values for a valid demographic
model (default is 1000)}
}
\description{
This is a core function for ABC inference using demografr. It generates simulation
replicates and computes summary statistic for the next step of an inference procedure,
which is the ABC estimation itself.
}
