% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_grid.R
\name{simulate_grid}
\alias{simulate_grid}
\title{Simulate values of summary statistics across a parameter grid}
\usage{
simulate_grid(
  model,
  grid,
  functions,
  replicates,
  sequence_length,
  recombination_rate,
  mutation_rate = 0,
  split = FALSE,
  model_args = NULL,
  engine_args = NULL,
  packages = NULL,
  engine = c("msprime", "slim"),
  file = NULL
)
}
\arguments{
\item{model}{A slendr model generating function}

\item{grid}{A data frame object containing parameter grid such as one produced by
\code{tidyr::expand_grid} or \code{base::expand.grid}}

\item{functions}{A named list of summary statistic functions to apply on simulated
tree sequences}

\item{replicates}{How many simulation replicates to run for each parameter combination?}

\item{sequence_length}{Amount of sequence to simulate using slendr (in numbers of basepairs)}

\item{recombination_rate}{Recombination rate to use for the simulation}

\item{mutation_rate}{Mutation rate to use for the simulation}

\item{model_args}{Optional (non-prior) arguments for the scaffold model generating function}

\item{engine_args}{Optional arguments for the slendr simulation back ends}

\item{packages}{A character vector with package names used by user-defined summary statistic
functions. Only relevant when parallelization is set up using \code{future::plan()} to make
sure that the parallelized tree-sequence summary statistic functions have all of their
packages available.}

\item{engine}{Which simulation engine to use? Values "msprime" and "slendr::"
will use the built-in slendr simulation back ends.}

\item{file}{If not \code{NULL}, a path where to save the data frame with simulated grid results}
}
\value{
If \code{file != NULL}, returns a data frame with simulated grid results. Otherwise
  does not return anything, saving an object to an .rds file instead.
}
\description{
For a given data frame (each column a parameter of a slendr model function) simulates
values of given population genetic statistics
}
