% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.R
\name{simulate_model}
\alias{simulate_model}
\title{Simulate a single tree sequence from the given inference setup}
\usage{
simulate_model(
  model,
  parameters,
  sequence_length,
  recombination_rate,
  mutation_rate = 0,
  data = NULL,
  format = c("ts", "files"),
  engine = NULL,
  model_args = NULL,
  engine_args = NULL,
  attempts = 1000
)
}
\arguments{
\item{model}{Either a slendr model generating function (in which case \code{engine} must
be either "msprime" or "slim", i.e. one of the two of slendr's simulation back ends),
or a path to a custom user-defined SLiM or msprime script (in which case \code{engine}
must be "files").}

\item{parameters}{A list of prior distributions to use for sampling of model parameters}

\item{sequence_length}{Amount of sequence to simulate using slendr (in base pairs).
Ignored when custom simulations scripts are provided.}

\item{recombination_rate}{Recombination rate to use for the simulation
Ignored when custom simulations scripts are provided.}

\item{mutation_rate}{Mutation rate to use for the simulation
Ignored when custom simulations scripts are provided.}

\item{data}{A named list of data-generating functions. The names then represent all possible
arguments of simulated summary statistic functions.}

\item{format}{In which format will the model generate results to be used for computing
simulated summary statistics?}

\item{engine}{Which simulation engine to use? Values "msprime" and "slim" will use one of
the built-in slendr simulation back ends. Which engine will be used is determined
by the nature of the \code{model}. If \code{engine = NULL}, then spatial slendr models will
by default use the "slim" back end, non-spatial models will use the "msprime" back end, and
custom user-defined model scripts will use the "custom" engine. Setting this argument
explicitly will change the back ends (where appropriate). Setting this argument for custom
simulation script has no effect.}

\item{model_args}{Optional (non-prior) arguments for the slendr model generating function.
Setting this argument for custom simulation script has no effect.}

\item{engine_args}{Optional arguments for the slendr simulation back end. Setting this
argument for custom simulation script has no effect.}

\item{attempts}{Maximum number of attempts to generate prior values for a valid demographic
model (i.e. model which generates an output without error, default is 1000)}
}
\value{
Either a tree-sequence object of the class \code{slendr_ts} when a slendr model
  was simulated, or a path to an output file when a custom simulation script was used.
}
\description{
Simulates a tree sequence object from a model with parameters as specified either
by sampling from priors or by a given list of fixed parameter values
}
\details{
This function is useful to generate a small tree sequence to be used when developing
summary statistic functions inference using demografr.
}
\examples{
slendr::check_dependencies(python = TRUE, quit = TRUE)

library(demografr)

library(slendr)
init_env(quiet = TRUE)

##################################################
# define a model

model <- function(Ne_A, Ne_B, Ne_C, Ne_D, T_AB, T_BC, T_CD, gf_BC) {
  A <- population("A", time = 1,    N = Ne_A)
  B <- population("B", time = T_AB, N = Ne_B, parent = A)
  C <- population("C", time = T_BC, N = Ne_C, parent = B)
  D <- population("D", time = T_CD, N = Ne_D, parent = C)

  gf <- gene_flow(from = B, to = C, start = 9000, end = 9301, rate = gf_BC)

  model <- compile_model(
    populations = list(A, B, C, D), gene_flow = gf,
    generation_time = 1, simulation_length = 10000,
    direction = "forward", serialize = FALSE
  )

  samples <- schedule_sampling(
    model, times = 10000,
    list(A, 25), list(B, 25), list(C, 25), list(D, 25),
    strict = TRUE
  )

  # when a specific sampling schedule is to be used, both model and samples
  # must be returned by the function
  return(list(model, samples))
}

##################################################
# set up priors

priors <- list(
  Ne_A  ~ runif(1000, 3000),
  Ne_B  ~ runif(100,  1500),
  Ne_C  ~ runif(5000, 10000),
  Ne_D  ~ runif(2000, 7000),

  T_AB  ~ runif(1,    4000),
  T_BC  ~ runif(3000, 9000),
  T_CD  ~ runif(5000, 10000),

  gf_BC ~ runif(0, 0.3)
)

##################################################
# prepare a list of empirical summary statistics

observed_diversity <- read.table(system.file("examples/basics_diversity.tsv", package = "demografr"), header = TRUE)
observed_divergence <- read.table(system.file("examples/basics_divergence.tsv", package = "demografr"), header = TRUE)
observed_f4  <- read.table(system.file("examples/basics_f4.tsv", package = "demografr"), header = TRUE)
observed <- list(
  diversity  = observed_diversity,
  divergence = observed_divergence,
  f4         = observed_f4

)

##################################################
# prepare a list of simulated summary statistics

compute_diversity <- function(ts) {
  samples <- ts_names(ts, split = "pop")
  ts_diversity(ts, sample_sets = samples)
}
compute_divergence <- function(ts) {
  samples <- ts_names(ts, split = "pop")
  ts_divergence(ts, sample_sets = samples)
}
compute_f4 <- function(ts) {
  samples <- ts_names(ts, split = "pop")
  A <- samples["A"]; B <- samples["B"]
  C <- samples["C"]; D <- samples["D"]
  ts_f4(ts, A, B, C, D)
}

functions <- list(
  diversity  = compute_diversity,
  divergence = compute_divergence,
  f4         = compute_f4
)

##################################################
# simulate data from a single model run
# (step #1 of a single ABC replicate simulation)

ts <- simulate_model(model, priors, sequence_length = 1e6, recombination_rate = 1e-8)

##################################################
# simulate data from a single model run
# (step #2 of a single ABC replicate simulation)

summarise_data(ts, functions)

##################################################
# validate all components of the ABC inference pipeline

validate_abc(model, priors, functions, observed,
             sequence_length = 1e6, recombination_rate = 1e-8)

#
# we're skipping the remaining steps because they are extremely
# computationally intensive for the scope of this example
#

##################################################
# set up paralelization

# library(future)
# plan(multisession, workers = availableCores())

##################################################
# simulate data (summary statistics)

# data <- simulate_abc(
#   model, priors, functions, observed, iterations = 10000,
#   sequence_length = 10e6, recombination_rate = 1e-8, mutation_rate = 1e-8
# )

##################################################
# perform ABC inference

# abc <- run_abc(data, engine = "abc", tol = 0.01, method = "neuralnet")
}
