% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.R
\name{summarise_data}
\alias{summarise_data}
\title{Apply summary statistic functions to the simulated data}
\usage{
summarise_data(data, functions)
}
\arguments{
\item{data}{Data generated by \code{simulate_model}}

\item{functions}{A named list of data-generating functions}
}
\value{
A named list of data frames, one summary statistic for each
}
\description{
Apply summary statistic functions to the simulated data
}
