% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack}
\alias{unpack}
\title{Unpack demografr object into individual components of the abc package}
\usage{
unpack(object)
}
\arguments{
\item{object}{A demografr object to be decomposed to individual components}
}
\description{
Unpack demografr object into individual components of the abc package
}
