% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_abc.R
\name{validate_abc}
\alias{validate_abc}
\title{Validate individual components of an ABC model}
\usage{
validate_abc(
  model,
  priors,
  functions,
  observed,
  model_args = NULL,
  sequence_length = 10000,
  recombination_rate = 0,
  mutation_rate = 0,
  max_attempts = 1000,
  ...
)
}
\arguments{
\item{model}{A slendr model generating functions}

\item{priors}{A list of prior distributions to use for sampling of model parameters}

\item{functions}{A named list of summary statistic functions to apply on simulated
tree sequences}

\item{observed}{A named list of observed summary statistics}

\item{model_args}{Optional non-prior arguments for the model generating function
(specified as a named list)}

\item{sequence_length}{Amount of sequence to simulate using slendr (in numbers of basepairs)}

\item{recombination_rate}{Recombination rate to use for the simulation}

\item{mutation_rate}{Mutation rate to use for the simulation}

\item{max_attempts}{Maximum number of attempts to generate prior values for a valid demographic
model (default is 1000)}
}
\value{
No return value. The function is ran for its terminal output.
}
\description{
Validates the ABC setup by checking that all priors can be correctly sampled from,
that a slendr model resulting from those priors can simulate a tree sequence,
and that the user-defined summary functions produce output compatible with
the provided empirical summary statistics.
}
