% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_abc.R
\name{validate_abc}
\alias{validate_abc}
\title{Validate individual components of an ABC model}
\usage{
validate_abc(
  model,
  priors,
  functions,
  observed,
  sequence_length = 10000,
  recombination_rate = 0,
  mutation_rate = 0,
  format = c("ts", "files"),
  engine = NULL,
  model_args = NULL,
  engine_args = NULL,
  quiet = FALSE,
  attempts = 1000
)
}
\arguments{
\item{model}{Either a slendr model generating function (in which case \code{engine} must
be either "msprime" or "slim", i.e. one of the two of slendr's simulation back ends),
or a path to a custom user-defined SLiM or msprime script (in which case \code{engine}
must be "custom").}

\item{priors}{A list of prior distributions to use for sampling of model parameters}

\item{functions}{A named list of summary statistic functions to apply on different
simulated results (tree sequences or custom files)}

\item{observed}{A named list of observed summary statistics}

\item{sequence_length}{Amount of sequence to simulate using slendr (in base pairs).
Ignored when custom simulations scripts are provided.}

\item{recombination_rate}{Recombination rate to use for the simulation}

\item{mutation_rate}{Mutation rate to use for the simulation}

\item{engine}{Which simulation engine to use? Values "msprime" and "slim" will use one of
the built-in slendr simulation back ends. Which engine will be used is determined
by the nature of the \code{model}. If \code{engine = NULL}, then spatial slendr models will
by default use the "slim" back end, non-spatial models will use the "msprime" back end, and
custom user-defined model scripts will use the "custom" engine. Setting this argument
explicitly will change the back ends (where appropriate). Setting this argument for custom
simulation script has no effect.}

\item{model_args}{Optional (non-prior) arguments for the slendr model generating function.
Setting this argument for custom simulation script has no effect.}

\item{engine_args}{Optional arguments for the slendr simulation back end. Setting this
argument for custom simulation script has no effect.}

\item{quiet}{Should the log output of the validation be printed to the console?
(Default is \code{TRUE}.)}

\item{attempts}{Maximum number of attempts to generate prior values for a valid demographic
model (default is 1000)}
}
\value{
No return value. The function is ran for its terminal output.
}
\description{
Validates the ABC setup by checking that all priors can be correctly sampled from,
that a slendr model resulting from those priors can simulate a tree sequence,
and that the user-defined summary functions produce output compatible with
the provided empirical summary statistics.
}
